requirements.txt
django==5.1.3
djangorestframework==3.15.2
selenium==4.19.0
webdriver-manager==4.0.1
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
codecov==2.1.13
